requirements.txt
aiogram==2.22.1
aiohttp==3.8.1
aiosignal==1.2.0
APScheduler==3.9.1
async-timeout==4.0.2
attrs==22.1.0
Babel==2.9.1
cachetools==5.2.0
certifi==2022.9.14
charset-normalizer==2.1.1
DAWG-Python==0.7.2
docopt==0.6.2
docxcompose==1.3.5
docxtpl==0.16.4
frozenlist==1.3.1
fuzzywuzzy==0.18.0
google-api-core==2.10.1
google-api-python-client==2.61.0
google-auth==2.11.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.5.3
googleapis-common-protos==1.56.4
httplib2==0.20.4
idna==3.4
Jinja2==3.1.2
Levenshtein==0.20.8
lxml==4.9.1
MarkupSafe==2.1.1
multidict==6.0.2
oauthlib==3.2.1
Petrovich==2.0.1
protobuf==4.21.6
pyasn1==0.4.8
pyasn1-modules==0.2.8
pymorphy2==0.9.1
pymorphy2-dicts-ru==2.4.417127.4579844
pyparsing==3.0.9
python-docx==0.8.11
python-Levenshtein==0.20.8
pytz==2022.2.1
pytz-deprecation-shim==0.1.0.post0
rapidfuzz==2.13.2
requests==2.28.1
requests-oauthlib==1.3.1
rsa==4.9
six==1.16.0
tzdata==2022.2
tzlocal==4.2
uritemplate==4.1.1
urllib3==1.26.12
XlsxWriter==3.0.3
yarl==1.8.1
